InterruptContext is for dorks.

+++ src/constants/serial.rs
@@ -1,2 +1,2 @@
-pub const SERIAL_INTERRUPT: u32 = 0x24;
+pub const SERIAL_INTERRUPT: u8 = 0x24;
 pub const COM1: u16 = 0x3F8;
